Started work on a new UI.
Old UI is built on janky pixijs custom widgets.  pixijs is great for render surfaces, not so great for good interactive widgets.  Building something better around react for the controls, while still using pixi for the render surface.

Incidentally also added a units buffer and initialization for it to the backend.