UI for specifying tile generation ruleset.
Switched over to using jotai for state management as the
react primitives were getting too cumbersome.

Changed the public program representation to a text
assembly format that's processed by a server command, and
provides a really minimal parsing and error API that can be
used to do highlights of error lines, etc.